requirements/base.txt
aio-pika==9.5.7
aiofiles==24.1.0
aiormq==6.9.2
alembic==1.17.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
astroid==4.0.2
asyncpg==0.30.0
boto3==1.34.0
botocore==1.34.162
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
dill==0.4.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.1
fastapi-cache2==0.2.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
isort==7.0.0
jmespath==1.0.1
loguru==0.7.2
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
multidict==6.7.0
packaging==25.0
pamqp==3.3.0
pendulum==3.1.0
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pylint==4.0.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.1
redis==5.0.1
requests==2.32.5
ruff==0.14.5
s3transfer==0.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0